hotfix: Migrate the pipeline to native too calling agent and revert faulty mastery calculation
