+++ app/pipeline/shared/utils.py
@@ -1,0 +1,23 @@
+from typing import Callable, List
+
+from langchain_core.tools import StructuredTool
+
+
+def generate_structured_tool_from_function(tool_function: Callable) -> StructuredTool:
+    """
+    Generates a structured tool from a function
+    :param tool_function: The tool function
+    :return: The structured tool
+    """
+    return StructuredTool.from_function(tool_function)
+
+
+def generate_structured_tools_from_functions(
+    tools: List[Callable],
+) -> List[StructuredTool]:
+    """
+    Generates a list of structured tools from a list of functions
+    :param tools: The list of tool functions
+    :return: The list of structured tools
+    """
+    return [generate_structured_tool_from_function(_tool) for _tool in tools]
